chore: rename all occurrences of space invaders with i8080 arcade
