+++ README.md
@@ -1,7 +1,14 @@
 
 ### Introduction
 
-This demo project shows how to make use of the [mach-emu sdk](http://github.com/nbeddows/mach-emu/) and [meen_hw](http://github.com/nbeddows/meen-hw) to emulate an arcade machine, in this case, Space Invaders. I don't consider the emulation to be the most efficient or accurate, but I'm happy with where it is at.
+This demo project shows how to make use of the [mach-emu sdk](http://github.com/nbeddows/mach-emu/) and [meen_hw](http://github.com/nbeddows/meen-hw) to emulate an arcade machine, in this case, one based on the Space Invaders arcade hardware. I don't consider the emulation to be the most efficient or accurate, but I'm happy with where it is at.
+
+The project has been tested against the following roms:
+
+- Space Invaders.
+- Space Invaders Deluxe.
+- Balloon Bomber (This one looks to have issues which go beyond the superficial that require further investigation).
+- Lunar Rescue.
 
 ### Compilation
 
@@ -34,7 +41,7 @@ This project uses [CMake (minimum version 3.23)](https://cmake.org/) for its bui
 **NOTE**: when performing a cross compile using a host profile you must install the requisite toolchain of the target architecture, [see pre-requisites](#pre-requisites).
 
 **NOTE**: under Linux errors similar to the following, `ERROR: xorg/system: Error in system_requirements() method` require additional package installations as denoted by the above console messages: "`dpkg-query: no packages found matching ${pkg0}`": `sudo apt install ${pkg0} ${pkg1} ${pkgn}`.
-When cross compiling for arm you may need to add the arm development repositories to your apt sources if the packages previously installed could not be found, for example:
+When cross compiling for arm you may need to add the arm development repositories to your apt sources if the packages previously installed could not be found, for example (at the time of writing):
 - `sudo nano /etc/apt/source.list`
 - Append the following:
     - deb [arch=arm64] http://ports.ubuntu.com/ lunar main multiverse universe
@@ -61,24 +68,24 @@ The following dependent packages will be (compiled if required and) installed:
 - Multi configuration generators (MSVC for example): `cmake --preset conan-default [-Wno-dev]`.
 - Single configuration generators (make for example): `cmake --preset conan-release [-Wno-dev]`.
 
-**4.** Run cmake to compile Space Invaders: `cmake --build --preset conan-release`.
+**4.** Run cmake to compile i8080-arcade: `cmake --build --preset conan-release`.
 
-**5.** Run Space Invaders:
+**5.** Run i8080-arcade:
 
 **Linux/Windows (x86_64)**:
 - `build\generators\conanrun.bat/sh`: export the dependent shared library paths.
-- `artifacts/Release/x86_64/bin/space-invaders`.
+- `artifacts/Release/x86_64/bin/i8080-arcade`.
 - `build\generators\deactivate_conanrun.bat/sh`: restore the environment.
 
 **Linux (armv7hf, armv8)**:
 
 When running a cross compiled build the binaries need to be uploaded to the host machine before they can be executed.
 1. Create an Arm Linux binary distribution: See building a binary package. 
-2. Copy the distribution to the arm machine: `scp build/Release/Sdk/space-invaders-v0.5.0-Linux-armv7hf-bin.tar.gz ${user}@raspberrypi:space-invaders-v0.5.0.tar.gz`.
+2. Copy the distribution to the arm machine: `scp build/Release/Sdk/i8080-arcade-v0.6.0-Linux-armv7hf-bin.tar.gz ${user}@raspberrypi:i8080-arcade-v0.6.0.tar.gz`.
 3. Ssh into the arm machine: `ssh ${user}@raspberrypi`.
-4. Extract the space-invaders archive copied over via scp: `tar -xzf space-invaders-v0.5.0.tar.gz`.
-5. Change directory to space-invaders `cd space-invaders`.
-6. Run Space Invaders: `./run-space-invaders.sh`.<br>
+4. Extract the i8080-arcade archive copied over via scp: `tar -xzf i8080-arcade-v0.6.0.tar.gz`.
+5. Change directory to i8080-arcade `cd i8080-arcade`.
+6. Run i8080-arcade: `./run-i8080-arcade.sh`.<br>
 
 The following command line options are available:
 
@@ -93,9 +100,9 @@ The following command line options are available:
 
 A standalone binary package can be built via `cpack` that can be distributed and installed:
 
-- `cmake --build --preset conan-release --target=space-invaders-pkg`
+- `cmake --build --preset conan-release --target=i8080-arcade-pkg`
 
-The `space-invaders-pkg` target defined in the root CMakeLists.txt will build a tar gzipped package via the following cpack command:
+The `i8080-arcade-pkg` target defined in the root CMakeLists.txt will build a tar gzipped package via the following cpack command:
 
 - `cpack --config build\CPackConfig.cmake -C ${buildType} -G TGZ`
 
@@ -109,7 +116,7 @@ This will build a binary package using the `zip` utility.
 
 Run `cpack --help` for a list available generators.
 
-The `space-invaders-pkg` target will also strip all binary files where applicable and will also perform static analysis of the source code via `cppcheck`.
+The `i8080-arcade-pkg` target will also strip all binary files where applicable and will also perform static analysis of the source code via `cppcheck`.
 
 ### Configuration
 
@@ -124,7 +131,7 @@ These options should be fixed to the specfied values unless stated otherwise.
 The current settings for these options should be sufficient, changing them may have a negative impact on performance.
 
 `clockResolution:1000000000 / 60 / 2` - i8080 arcade hardware runs at 60Hz with 2 interrupts per frame, set the machine clock resolution accordingly.<br>
-`isrFreq:0.9` - We require 4 interrupts, 2 for Space Invaders and 2 machine level interrupts for loading and saving. Ideally we would lock the interrupt service routine frequency to the clock resolution ("isrFreq":1), however, we need to spare some time for checking for load and save requests, so we bump the isrFreq down by ten percent ("isrFreq":0.9). One could lower it further, this would make it more responsive (0.9 should be good enough). Increasing it above 1 would make it slower and not respond to load/save requests.<br>
+`isrFreq:0.9` - We require 4 interrupts, 2 for i8080-arcade and 2 machine level interrupts for loading and saving. Ideally we would lock the interrupt service routine frequency to the clock resolution ("isrFreq":1), however, we need to spare some time for checking for load and save requests, so we bump the isrFreq down by ten percent ("isrFreq":0.9). One could lower it further, this would make it more responsive (0.9 should be good enough). Increasing it above 1 would make it slower and not respond to load/save requests.<br>
 `loadAsync:true` - Load the machine state asynchronously.<br> 
 `runAsync:true` - Run the machine asynchronously from the io.<br>
 `saveAsync:true` - Save the machine state asynchronously.<br>
